change: - Additional constructor with boolean 'deleteOnExit'.
SVN: 2582

+++ common/sourceTest/java/ch/systemsx/cisd/common/utilities/AbstractFileSystemTestCase.java
@@ -40,8 +40,16 @@ public abstract class AbstractFileSystemTestCase
 
     protected AbstractFileSystemTestCase()
     {
+        this(true);
+    }
+
+    protected AbstractFileSystemTestCase(final boolean deleteOnExit)
+    {
         workingDirectory = new File(UNIT_TEST_ROOT_DIRECTORY, getClass().getSimpleName());
-        workingDirectory.deleteOnExit();
+        if (deleteOnExit)
+        {
+            workingDirectory.deleteOnExit();
+        }
         LogInitializer.init();
     }
